fix: flat cubes and correct color of cubes
