+++ app/pages/home/index.tsx
@@ -1,4 +1,4 @@
-import React, { Suspense, useRef, useState } from 'react';
+import React, { useRef, useState } from 'react';
 import { View } from 'react-native';
 import { Color, PerspectiveCamera, Vector3 } from 'three';
 import { Canvas } from '@react-three/fiber/native';
@@ -7,7 +7,7 @@ import { homePageStyles } from './style';
 import FixedControlCameraButtons from './component';
 import Cube from '../../component/cube';
 import { getRotateCamera } from './controlCamera';
-import { MineCraftGround } from '../../utils/buildScene';
+import { MineCraftGround } from '../../component/mineCraftGround';
 
 export default function HomePage(): React.JSX.Element {
   const cameraRef = useRef<PerspectiveCamera>(new PerspectiveCamera(60, window.innerWidth / window.innerHeight, 1, 20000));
@@ -23,15 +23,14 @@ export default function HomePage(): React.JSX.Element {
         style={homePageStyles.canvasView}
         {...rotateCameraEvents}
       >
-        <Canvas scene={{ background: new Color(0xbfd1e5) }} camera={camera}>
-          <Suspense fallback={null}>
-            <Cube position={[1, 1, 1]} />
-            <XYZGrid isXYZ="isX" />
-            <XYZGrid isXYZ="isY" />
-            <XYZGrid isXYZ="isZ" />
-            <MineCraftGround />
-            <directionalLight args={[0xffffff, 12]} position={new Vector3(1, 1, 0.5).normalize()}/>
-          </Suspense>
+        <Canvas scene={{ background: new Color(0xbfd1e5)  }} camera={camera} >
+          <Cube position={[1, 1, 1]} />
+          <XYZGrid isXYZ="isX" />
+          <XYZGrid isXYZ="isY" />
+          <XYZGrid isXYZ="isZ" />
+          <MineCraftGround />
+          <ambientLight color={0xeeeeee} intensity={3} />
+          <directionalLight color={0xffffff} intensity={12} position={new Vector3(1, 1, 0.5).normalize()}/>
         </Canvas>
       </View>
     </View>
